+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - UFO Abduction
 - Supersonic
 - UFO Abducts Zombies
+- Roll Dice
 
 [1.1.1]
 
@@ -61,6 +62,7 @@
 - Effect "Teleport To Nearest Basement": now finds better square to teleport
 - Effects "Launch Player Up" and "Launch Everyone Up" now do special actions with custom damage
 - Effects "Add Bomb To Player Inventory", "Random Item Bomb", and "Spawn Explosive Spiffos" now show a progress bar countdown to detonation via UI
+- Effect "Medieval Times" now shows a green "Modified N zombies" chat line on the player when it adds helmets
 
 [1.1.0]